Types & Helpers
The helpers for 'Atom' and such have been put into 'types.c',
with their declarations in 'types.h'. This allows them to be
used throughout the program easily.

Also introduce the idea of the 'BuiltIn' atom. BuiltIns are
exactly what they sound like; built in C functions that can
be called from the lisp environment with a particular binding.