Better `+` and `*` operators
The `+` and `*` operators now accept any amount of
integers as arguments, instead of only two.

This is all thanks to let bindings!
They really change what is possible in LISP.

The documentation was also slightly restructured,
as most renderings include the title as a top-level header.

Also a few new builtins, as well as the `ENV` operator.

The first two debug flags were introduced, being
`DEBUG/MACRO` and `DEBUG/ENVIRONMENT`.