Generalize docstring to `DEFINE`
`DEFINE` can now handle all docstring capabilities.
`LAMBDA` no longer needs special handling of a docstring,
as an unbound function wouldn't ever be able to have it's
docstring gotten in the first place.
`MACRO` still needs to handle a docstring explicitly,
as it doesn't ever use a `DEFINE`; in fact, it now
*requires* a docstring, so as to make each macro well-documented.
I may make it possible to pass NIL as a docstring in the future.