Continuation
Basically, all of the execution state of the program is saved
in the stack frame, and the stack is passed around by pointer.

This effectively removes the language stack into `malloc()` territory,
meaning there won't be an actual stack overflow, given enough
recursion.

The following example would crash due to a stack overflow,
but now works flawlessly (given you have enough free RAM).

``` lisp
(defun count (n)
  "Count up to N."
  (if (= n 0)
    0
    (+ 1 (count (- n 1)))))
```

This language does still leak memory,
but we are 90% of the way to fixing that.