Being Good Bois
Memory safety in C is a hot-button issue, and is always
a main problem in software written in this language.

With this commit, we hope to finally be good bois;
that is, for (almost) all the memory we allocate,
it gets freed before the program terminates.

To ensure this happens, counters have been added that
correspond to how many allocations have been made
vs how many have been freed. I think it would
be cool to make this data available in LISP
so that maybe someone could make a plugin/script
that displays memory info or something.
That's an issue for another commit :^).

Changes have been made to the garbage collector so
as to not attempt to free the same pair twice.
This is done by removing a freed pair from the
global allocations list (prev->next = this->next).