Atom StringBuilder
While it isn't quite as easy as, say, using Java, it is still
possible to build dynamic strings in standard C. There are
likely bugs in this implementation, and it is by no means perfect,
but it is the start of something much larger.

Using this new string building function, we may construct docstrings
that include the arguments for a given closure, for example. This is
very useful to the end user, to see the closure's call signature.

It is known that the strings generated from lists are not pretty.