Remove `GenericAllocation`
It was a mess, anyway. Plus, only like a total of 20 would
ever be allocated/freed in an average LITE LISP session.

Obviously, with the removal of this garbage collection,
garbage now builds up concerning strings. As of now,
this shouldn't matter, but it very quickly will.

A solution to this memory allocation problem is necessary.