Make a (graphical) mess
This setup should provide a somewhat easy way to switch out
graphical backends, as well as give LITE a GUI in general.

The initial backend chosen is SDL, as it is cross
platform itself and gets us a lot of ground quickly.

From here, we need to make `GUIInput` usable, as it
currently is a joke from a usefulness perspective.

Also, make the docs better!!!