GFX: Two-way communication
This allows for much better input handling (I believe).

While SDL does have TextInput events now, I think the
best thing to keep this library backend-agnostic is to
handle single characters coming in at a time in LITE itself.
Honestly, I don't know if this is the right solution, but progress
is progress and we need to move forward to avoid any stagnation.