GFX: Include SDL2.dll
This makes it easier to get LITE GFX up and running on Windows,
as no DLLs are required to be copy and pasted from SDL2 bin.

It may be worth it to eventually store these with the SDL2 backend,
and only copy them to the bin folder when the backend is used.