¡Association lists!
Association lists, better known as alists, /associate/ one value
to another value. Think of it like a LISP dictionary type.

We also use the API right away for shift-remapping that may
be altered from the LISP environment (once we make the proper
LISP functions and/or macros in the LITE LISP std library).