Fix `alist_get()` crashing program
This was caused by de-referencing a bad pointer through `list_get()`.
To fix it, we must keep track of the rest of the list as we go to
ensure the next item in the list is non-nil before moving on.

Also introduced is `alistp()`, which further ensures
the input is valid in `alist_get()` and `alist_set()`.

The SDL backend was updated very slightly with refactoring to print
minimal output in while loops when context has `NULL` pointers.
Also, non-fatal errors are no longer treated as fatal.