Move `SYM` from special form to BuiltIn
`SYM` doesn't require any special knowledge, so it can be delegated to
a BuiltIn. `ENV` requires the special form due to needing the current
LISP evaluation environment, which can only be provided from within
the LISP evaluation procedure.

Also, that idea in the comment to return symbols from the environment
with `ENV`, that's literally what symbols do all on their own `:^O`.