Cross-compatibility changes
Turns out I used a Windows-only type-alias in the SDL backend,
so that has been fixed up. There was also an implicit include
in `rope.c` that is now explicitly included :^).

Also added an example of how to install dependencies on Linux
since the website is kind of lacking in that regard.