LITE LISP: Implement `while` loops
It was kind of rough figuring out how the continuation would work,
but I figured it out after a few hours of flubbing about.

I believe the handling of the body evaluation could be much better,
but at least it only recurses the amount of nested while loops that
there are (if that makes any sense).