¡Keymaps!
It is true! LITE GFX now responds to a LISP keymap!
This is very cool, as the input can be used to alter
the input, which can be used to... you get the idea.

This will especially be useful to bind `save-file`, `switch-buffer`,
and so much more! I am excited to get this working as a real editor!