Evaluating buffer contents
We can now evaluate arbitrary LISP strings as LISP, as well as get a
line, set of lines, the current line, or even the entire buffer as a
string. This allows LITE buffers to be evaluated as LITE LISP!

We are edging closer and closer to a usable, extensible text editor.