Hunt down heap corruption bug
Turns out that `smpl.path = path` caused me to lose hours of my life,
that I will never get back. The problem lies in the fact that the path
is freed upon freeing the file, which with simple files is right after
use, usually. This meant that the path to the file was being freed,
and the OS was re-using that heap memory when it could (meaning the
code worked until the OS decided it didn't, different every time).

To fix it, we can simply `strdup()` the path into the simple file.

I ended up making lots and lots of changes while hunting this bug
down, and the ones that don't seem too egregious I've kept.