Fix off-by-one error in `Rope` implementation
With this also comes underflow protection on subtraction of unsigned
integers in `rope_remove_span()`. This should never be needed due to
the other changes made in this commit, but it will help protect from
future changes possibly breaking the rope implementation.

Also prefer `memcpy` over `strncpy`, writing zeros is something we do
ourselves most of the time.