¡Prompt for user input from LISP!
This was actually a really difficult problem to solve for me for some
reason. I am glad to have made this much progress, though, as LITE
LISP may now get user input as a LISP string object! This is very
cool, and allows for even more interaction between the user and the
LISP environment.

This works for both the REPL and the graphical version of LITE.