Rebind `\r` input to `\n`
SDL passes return as a carriage return character vs a newline.
This meant that our files had line endings of a single `\r`,
which is actually kind of terrifying. This rebinds the `\r`
character to `\n`, giving the more traditional linefeed newline.

We really need to convert to string-based input rather than
character-based!

Also in this commit:
- Open files for reading in binary mode to prevent byte-count/
character-count mismatches.
- Display "Saved." message after saving current buffer.