Make it possible to close LITE from anywhere
Previously, the close event wasn't handled properly when reading user
input. This is now fixed, due to a more robust GUI implementation as
well as an abstraction to it's own file.

This marks the first commit where doxygen-like comment syntax has been
introduced. I don't dislike it, and it is cool to have docs automatic.