Parser Upgrade: Handle basic escape sequences
We now handle two different escape sequences within strings in LITE
LISP: "\\r" and "\\n". The double-backslash tradition has been kept,
but that could easily change at any point in the future with search
and replace and a few lines of code changed in the parser.

Also, the buffer navigation in the standard library no longer relies
on the source files newline encoding! Cool!