No more (accidental) memory leaks
The program was run with Valgrind on Linux in a VM, and memory leaks
that were fixable were fixed. SDL2 has some leaks that don't grow over
time, so it's impossible to get zero errors with Valgrind, sadly.

At this point, the only leaks coming from LITE are for docstrings.
This is a known issue, and doesn't cause issues yet, so it's fine.