Standard Library: Another buffer navigation update
Instead of matching the newline bytes as a string, each byte is
checked individually.

Allow `progn` style body within `when` and `unless` (evaluate all
expressions in a row and return the last returned value).

`backward-line` is now much, much simpler (and shorter).

Also simplify `forward-line`. This fixes some bugs concerning multiple
newlines at the end of a buffer, among other things.

`progn` has also been introduced. I don't like the name, but I can't
really think of a short, descriptive name for it other than the
traditional one.

Also, `DEFINE` special form now acts on the global environment instead
of the local one. I'm not sure if this is what I want, but it's
necessary for things like `switch-current-buffer`, for example.
It also allows define to be used within a let binding and the value
sticking around, rather than just being thrown away in the hidden
lambda. I think a better solution would be to have a special form
for global bindings and one for local. Maybe use `set` for this?