Much, much better keybindings with `SET`
The reason keybindings were so cumbersome and ugly before is that we
had to keep things at the global scope to allow `define` to actually
take effect in the global environment. Now, with `set`, we can do that
explicitly from any environment. Now, we can use `let` to bind keymaps
and work with them in a more natural way.

Also, add "<shift> <enter>" keybind to insert DOS newline. This allows
files with any line endings to be edited with minimal extraneous
effort. Eventually, we'll want automatic handling of newlines based on
what exists in the file already, but we aren't really at that level
of complexity, yet :^).