New BuiltIn: `COPY`
Copy allows an atom to be duplicated. This ability allows the standard
library structures implementation to now get a `MAKE` macro, which
copies the given structure bound to symbol and returns it.
This allows for a single definition of a structure with multiple
implementations (for example, a vector3 could be defined and used
all over the plave for coordinates, math, etc). Very cool!