New BuiltIn: `MEMBER`
Member returns T iff the given element is an element of the given
list. This allows, for example, `SET-MEMBER` to check if the
id symbol it is attempting to bind is a member of the structure.

Also abstract `builtin_eq` functionality to `compare_atoms` so it
can be used in other places throughout the code-base.