¡Huge Keymap Update!
Key handling now uses strings, and with that we are able to expand the
abilities of LITE LISP quite heavily. It expands them so much that we
don't need any keymap code within the C source anymore, and it's
entirely handled from LITE LISP. Cool! :^)

Parsing of escape sequences within strings was improved, and we can
now even use multi-line strings using it!