BugFix: Heap corruption (off-by-one argument off-by-one)
As it turns out, the size arguments to malloc and memset must not be
identical! Memset argument must be one less than the malloc argument,
to ensure that the heap buffer is not over-run.

Also in this commit are changes that allow LITE to be compiled with
non-free, proprietary build tools that I choose not to name, but
are overwhelmingly popular in the modern world (although slowly
being phased out by truly free software, like that of GNU and FSF).