Initializing...
We now look for a user's home path in the environment with `getenv()`,
and, if found, attempt to load `.lite` from the directory.
This allows automatic configuration at startup, with no command line.

I am thinking about moving to a directory-based approach, with a
basic installation script that copies std library and such to home
path, as well as creates a basic configuration with some helpful
examples on how to get help :^).