New BuiltIn: `%`
This acts as it does in the C language, returning the remainder after
a division. Eventually we will implement a true Gaussian modulo. With
it, we can define a recursive greatest common denominator function.

I also implemented the Newtonian square root approximation.

Both functions seen in this paper by the MIT AI research lab (1960):
https://web.archive.org/web/20100717111134/http://history.siam.org/sup/Fox_1960_LISP.pdf

Also add a corresponding hint to the error that happens when passing
too few or too many arguments.