It's been a `WHILE`
`WHILE` was a right mess... Recursion out the wazoo (read: stack
overflows). This has been fixed using continuations as a design
pattern.

The control flow now goes:

1. Encounter WHILE, generate WHILE stack then evaluate condition within
   it. Result value handled in `evaluate_return_value()`.

2. Decide whether or not to continue the loop based on conditional
   result.

3. If condition is false, pop WHILE stack and continue evaluation.
   Otherwise, generate WHILE-BODY stack to evaluate WHILE-BODY.
   When that stack's body is done evaluating, pop back into WHILE
   stack, setting expression to evaluate condition (re-enter step 2).

And it works! We now have working while loops in LITE LISP that don't
eat the stack like the cookie monster eats cookies :^). This allows us
to implement `DOLIST`, a very handy macro when it comes to LISP lists.