BugFixes: `READ-PROMPTED`, return keystring, and more
The rope implementation had a bug where `rope_insert()` would call
`rope_append`, completing the action, but it would never actually
return the pointer that it returned, meaning it appeared to fail.

The string escape sequence of the parser now handles lots of
backslashes in a row much better, only treating it as an escape
sequence when a character other than a backslash is encountered.

`READ-PROMPTED` used the old character API for text input, and
therefore had the incorrect return keystring of "\r".

Add `EVALUATE-PROMPTED` to evaluate string input as LITE LISP.