Simplify
Evaluation of `WHILE` and `PROGN` has been simplified, in that dead
code has been removed (bad conditionals). They should actually
function properly now!

The standard library got an update, mainly to decrease startup time
before we get LISP compilation up and running. It seems like a good
idea to check for duplicates, but when defining a new keymap, there
are guaranteed to be none, so that precaution doesn't have to take
place. However, when the user sets a key in the keymap, it is probably
best to use `KEYBIND-SET` instead of `KEYBIND-PUSH`, to avoid
duplicates, and the keymap growing un-necessarily large.
Although it would be fine with single keys, setting nested keymaps
using push would be a bad idea, as it would copy the entire keymap.