BugFix: `ADD-TO-ALIST-IMPL`
As it turns out, this function was ill-formed before. I'm not exactly
sure why, but it caused errors when chaining calls to itself, i.e.:
`(add-to-alist-impl (add-to-alist-impl some-alist "key" 69) "key" 420)`

All is fixed now! Also, keymaps got a bit of a semantics update.
When defining a new key that is not a nested keymap, use keymap-push
for a fast insertion, as we don't have to worry about duplicating a
single key in a keymap (at least right now). However, when defining
a key that is bound to a nested keymap, use keymap-set to prevent
the nested keymap from being duplicated in the parent, as this is
no good for memory usage or keeping iteration speed fast.
Finally, if defining a key in the global keymap, use
set-bind-in-keymap.

All of these together keep keymap initialization fast while keeping
memory usage and needless growth down. Hurray!

I think it may be worth it to implement a hash map, either through
BuiltIns from C, in LITE LISP, or both.