¡New Parser!
It's a new parser! Sort of. The parsing of the primitives stayed
the same, but the overall parsing of an expression has moved from
recursion-based to continuation-based. This means that the stack
will not keep growing no matter how many nested lists you are trying to
parse. Obviously, this is a great thing! It should also increase
performance, but nothing was measured.