¡Region Highlighting!
With this commit, we see the first time we are able to actually
select bits of text more than a single byte! This is exciting, as
it looks and feels more and more like a text editor as we go!

`GUIColour` has been aliased to `GUIColor` in `gui.h`.

In the GFX::SDL2 backend, a use-after-free bug has been fixed.

The default background has been adjusted, to better contrast
between text, the cursor, and the region (selected text).