Clear buffer mark upon editing
Any time a buffer is edited, the mark is de-activated. This is the
standard behaviour of how text selection works, as the byte offset
moving around generally confuses the user.

Also in this commit:

- Fix Windows builds failing. `_LCD` family of functions don't seem to
be available in Windows version of SDL_ttf 2.20.0, or something. I'm
not able to get it to link properly.

- Optimize text properties. No longer do we cache draw position of
every single character every single frame! We now just get the draw
positions we need as we need them.

- `HOTFUNCTION` pre-processor directive moved to utility header.