¡Scripts!
You can now run LITE in script mode with the `--script` command line
flag. This flag causes LITE to exit after treating the rest of the
command line arguments as LITE LISP source filepaths and evaluating
them. This is going to be extremely useful for creating end-to-end
tests, as the output may be redirected and easily compared.

We also removed the environment being re-created message, as it got
in the way as well.