¡Tests!
It's official, we have end-to-end tests for LITE LISP now!!

This is really exciting :^).

Plus, I wrote it in one of the oldest programming languages: `Pascal`.
The weird part is it's really ergonomic and kind of reminds me of
Python but with an upgraded standard library. I like it a lot.

Either way, I have only written a few very basic tests, but I'm sure
that will change over time :^).