Improve Test Runner
The test runner can now be run from the root of the repository OR from
within the `tst` subdirectory. This is helpful because that is where
the executable is actually generated.

Also, make the print built-in use the standard print function, as it
would be near impossible to write tests with pretty-printing being
the only option.