Three new BuiltIns regarding graphical windows
We can now get the size of the window, change the size of the window,
as well as set fullscreen/windowed mode. This is great, as it allows
the user to set a custom resolution and such upon startup config.

We also had to move the GUI initialization back when not in script
mode, so that evaluated files at startup don't have access to an
uninitialized graphical library.