`ACCESS` macro for structures
The LITE LISP standard library now contains a macro that binds the
member values of a structure to their corresponding IDs. It's like
`LET` but for all of a structure's members, you know? This is quite
an ergonomic way of accessing structures.

Also update all the type comparisons in `evaluation.c` to their proper
macros for better readability and code consistency.