Attempt to load standard library by default
This will only work if the distribution is properly packaged with the
necessary resources, but will save from having to run from terminal.

It may be very, very smart to include some way to not automatically
load the standard library without having to remove it from it's place
in the filesystem.