¡Quitting!
Yay! The user can now stop execution of LISP, which is a rather handy
feature to have. It's not an ideal API or anything right now, but just
having it implemented is much more important that getting it
perfect. At least now the user can cancel a 'READ-PROMPTED', for
example.

Changes have not been tested in the REPL or file-based evaluation, so
results there may vary, and will likely need slight tweaks to get it
to work just right.