¡Symbol Table Overhaul!
The symbol table is no longer a linked list implemented through a
recursive pair; it is now a hash table! With this, we now spend an
order of magnitude less time (~%30 self-CPU to ~%0 self-CPU) just
doing symbol lookup, which is a huge plus. Now we have to tackle
`env_get()`...