Quit Optimization
As it turns out, calling env_get after literally any and every
expression is evaluated is a bit performance heavy. This was what we
were doing in order to handle the user quitting; now, it is handled
through a built-in which sets an internal global flag. This global
flag is then checked after each expression is evaluated, meaning it is
a simple boolean branch situation (branch-if-zero). This is so much
more efficient, and works just the same. There may still be bugs to
iron-out regarding the REPL, but LITE GFX is working 100% as expected.