`realpath()` is *not* like `_fullpath()`, sadly
Missing semicolon, haha.

It's not the 80s anymore: no more `get_current_dir_name()`

Error printout for `realpath()`

These stupid bikeshedding commits are needed so I can test it on my
linux VM by pulling from git; it will be fixed shortly (or not).

Just going to add: the fact that `realpath` *requires* a file to exist
at path just to give you an absolute path is bonkers and annoying.

If `realpath()` fails with ENOENT, create file and try again.

Typos

I forget C is case-sensitive, sometimes...

The millionth commit that solves nothing: More `realpath()`

This commit has been rebased and is a combination of 7 commits.