Remove windows with a non-integer ID after applying
This allows a text property that shouldn't stick around to easily and
automatically delete itself, while also not having to worry about the
whole integer/namespace thing when creating and maintaining a text
property.

We *should* still, at some point, make it so that the user can request
an unused ID for a text property in a given window, but it's not
necessary for right now, I don't think.