Docstrings are now moving towards being more usable
Before, the builtin was just completely broken with incorrect
typechecking, so there's that. Now, the docstring builtin does
do okay, but the `(ENV)` special form is all sorts of messed up
when trying to use the `LIST-FUNCTIONS` we wrote a while back;
I've commented out the code that no longer works, and added a
framework for code that will work once `(ENV)` does (if ever).

Also fix a cross-platform issue with MSVC and _Static_assert not being
usable (by just using `assert`).