[BuiltIns] Return Error instead of error type (huge!)
This means errors can now be effectively bubbled up even through
builtins, which is really nice, as if an error occurs within
`evaluate-string`, it can just return it and whatever called evaluate
string will report the error. One of the best parts is that you can
attach atoms to errors, and so `symbol not bound` errors now show the
user what symbol isn't bound, which is very nice.