[Refactor] Buffers now report back how many bytes were handled
This allows for builtins to return this as well, which allows
for certain navigation functions to take different actions depending
on whether or not the point was actually *able* to be moved, or if
bytes were actually able to be removed.

I can definitely see myself needing to do this for the buffer insert
functions as well, I just didn't need them here and so it wasn't done.