[CMake] The Build part of integrating tree-sitter
It's that easy! We may want to statically link with known common
language grammars, that way the user is not required to have them; we
could just as easily have an option to add known languages to the
build step, and a bunch of grammars will be built along with LITE.

Lots of options.