[GFX] Tree sitter integration refactor
It's much better now, and doesn't constantly load and unload the
dynamic library. In fact, it's even almost nearly somewhat
usable (it's quite slow still) for C!!! C!!!! If we do an auto-indent,
we'll basically have a full on code editor. POG! For simpler
languages, with simple grammars (like `un`), it goes really well, and
the slowdown *is* felt but not anything that would be a true
annoyance; Emacs does the same thing from time to time.