[GC] Implement explicit garbage collection frames
Basically, the mark and sweep allocator has been updated to support
nested evaluations that do not perform in the same environments or
share much of anything, really. This is normally an issue, as to
garbage collect, *everything* that must not be freed must also be
marked; that's hard to do when it isn't reachable from the current
evaluation stack or environment.

The update includes each explicit marking (which will not be undone
next GC but instead must be manually undone) getting it's own unique
marker. Then, when explicitly unmarking, only allocations marked with
the unique marker are unmarked. Explicit allocation marking must not
overwrite other explicit allocation marks; if one is encountered, the
entire tree should be skipped, as it will be freed some other time by
whoever explicitly marked it in the first place.

Possibly identified bug in `evaluate_next_expression()` that has to do
with popping evaluation stack too early?