[GFX] Keep cursor on screen when navigating
This *does* introduce a slight slowness to navigation, but it's
ultimately worth it to be able to properly move around a text file.

Things to work on still: inserting off the edges doesn't scroll,
because the internal buffer API alters point without a chance to ever
call this, or the oppurtunity to access the data needed. So, somehow,
we will need a "on buffer changed" hook or something to be able to
properly update the cursor on insertion. OR we *could* hard-code it in
the keymap handling so that when it inserts something, it keeps cursor
on screen from there. Lots of options.