[GFX/GL] More MSWin specific shenanigans
Copy DLLs to `bin` subdirectory, where executable is placed (or should
be). The user will still have to copy these to `Debug` or `Release`
subdirectories for MSVC, but it beats having to hunt these DLLs down
in the build tree.

Use proper IMPLIB on non-MSVC windows builds, as well.

Also simplify CMakeLists.txt by making a helper for long paths