[GFX/GL] Update screen on repeats
This is definitely not the right thing to do, but multiple repeat
events of, say, `ctrl-n` will cause it to block and not redraw until
the user lets go (not ideal for navigation). This re-entrance of the
loop allows for proper redrawing of cursor properties. Really, it'd be
ideal if we could just call some new API for the backend that will
re-update the GUIWindows from LISP and do properties and such, and
then we can just call `draw_gui` or something similar.